+++ manifest/caches/serializers.py
@@ -77,14 +77,15 @@ class NumpyByteSerializer(Serializer):
         Returns:
             normalized key.
         """
+        sub_response = response["response"]
         # Assume response is a dict with keys "choices" -> List dicts
         # with keys "array".
-        choices = response["choices"]
+        choices = sub_response["choices"]
         # We don't want to modify the response in place
         # but we want to avoid calling deepcopy on an array
-        del response["choices"]
-        response_copy = response.copy()
-        response["choices"] = choices
+        del sub_response["choices"]
+        response_copy = sub_response.copy()
+        sub_response["choices"] = choices
         response_copy["choices"] = []
         for choice in choices:
             if "array" not in choice:
@@ -101,7 +102,8 @@ class NumpyByteSerializer(Serializer):
                 hash_str = f.getvalue().hex()
             new_choice["array"] = hash_str
             response_copy["choices"].append(new_choice)
-        return json.dumps(response_copy, sort_keys=True)
+        response["response"] = response_copy
+        return json.dumps(response, sort_keys=True)
 
     def key_to_response(self, key: str) -> Dict:
         """
@@ -114,7 +116,7 @@ class NumpyByteSerializer(Serializer):
             unnormalized response dict.
         """
         response = json.loads(key)
-        for choice in response["choices"]:
+        for choice in response["response"]["choices"]:
             hash_str = choice["array"]
             byte_str = bytes.fromhex(hash_str)
             with io.BytesIO(byte_str) as f:
@@ -152,14 +154,15 @@ class ArraySerializer(Serializer):
         Returns:
             normalized key.
         """
+        sub_response = response["response"]
         # Assume response is a dict with keys "choices" -> List dicts
         # with keys "array".
-        choices = response["choices"]
+        choices = sub_response["choices"]
         # We don't want to modify the response in place
         # but we want to avoid calling deepcopy on an array
-        del response["choices"]
-        response_copy = response.copy()
-        response["choices"] = choices
+        del sub_response["choices"]
+        response_copy = sub_response.copy()
+        sub_response["choices"] = choices
         response_copy["choices"] = []
         for choice in choices:
             if "array" not in choice:
@@ -179,7 +182,8 @@ class ArraySerializer(Serializer):
             response_copy["choices"].append(new_choice)
             if not self.writer.contains_key(hash_str):
                 self.writer.put(hash_str, arr)
-        return json.dumps(response_copy, sort_keys=True)
+        response["response"] = response_copy
+        return json.dumps(response, sort_keys=True)
 
     def key_to_response(self, key: str) -> Dict:
         """
@@ -194,7 +198,7 @@ class ArraySerializer(Serializer):
             unnormalized response dict.
         """
         response = json.loads(key)
-        for choice in response["choices"]:
+        for choice in response["response"]["choices"]:
             hash_str = choice["array"]
             choice["array"] = self.writer.get(hash_str)
         return response