+++ manifest/clients/cohere.py
@@ -93,7 +93,7 @@ class CohereClient(Client):
         """
         return {"model_name": self.NAME, "engine": getattr(self, "engine")}
 
-    def format_response(self, response: Dict, request: Dict) -> Dict[str, Any]:
+    def validate_response(self, response: Dict, request: Dict) -> Dict[str, Any]:
         """
         Format response to dict.