+++ tests/test_client.py
@@ -33,10 +33,13 @@ def test_get_request() -> None:
         "prompt": "hello",
         "num_results": 3,
         "engine": "dummy",
+        "request_cls": "LMRequest",
     }
     assert response.get_json_response() == {
-        "choices": [{"text": "hello"}] * 3,
-        "usage": [{"prompt_tokens": 1, "completion_tokens": 1, "total_tokens": 2}] * 3,
+        "choices": [{"text": "hello", "token_logprobs": None, "tokens": None}] * 3,
+    }
+    assert response.get_usage_obj().dict() == {
+        "usages": [{"prompt_tokens": 1, "completion_tokens": 1, "total_tokens": 2}] * 3,
     }
 
     request_params = client.get_request("hello", {"n": 5})
@@ -45,10 +48,13 @@ def test_get_request() -> None:
         "prompt": "hello",
         "num_results": 5,
         "engine": "dummy",
+        "request_cls": "LMRequest",
     }
     assert response.get_json_response() == {
-        "choices": [{"text": "hello"}] * 5,
-        "usage": [{"prompt_tokens": 1, "completion_tokens": 1, "total_tokens": 2}] * 5,
+        "choices": [{"text": "hello", "token_logprobs": None, "tokens": None}] * 5,
+    }
+    assert response.get_usage_obj().dict() == {
+        "usages": [{"prompt_tokens": 1, "completion_tokens": 1, "total_tokens": 2}] * 5,
     }
 
     request_params = client.get_request(["hello"] * 5, {"n": 1})
@@ -57,8 +63,11 @@ def test_get_request() -> None:
         "prompt": ["hello"] * 5,
         "num_results": 1,
         "engine": "dummy",
+        "request_cls": "LMRequest",
     }
     assert response.get_json_response() == {
-        "choices": [{"text": "hello"}] * 5,
-        "usage": [{"prompt_tokens": 1, "completion_tokens": 1, "total_tokens": 2}] * 5,
+        "choices": [{"text": "hello", "token_logprobs": None, "tokens": None}] * 5,
+    }
+    assert response.get_usage_obj().dict() == {
+        "usages": [{"prompt_tokens": 1, "completion_tokens": 1, "total_tokens": 2}] * 5,
     }