+++ tests/test_serializer.py
@@ -10,23 +10,23 @@ def test_response_to_key_array() -> None:
     """Test array serializer initialization."""
     serializer = ArraySerializer()
     arr = np.random.rand(4, 4)
-    res = {"choices": [{"array": arr}]}
+    res = {"response": {"choices": [{"array": arr}]}}
     key = serializer.response_to_key(res)
     key_dct = json.loads(key)
-    assert isinstance(key_dct["choices"][0]["array"], str)
+    assert isinstance(key_dct["response"]["choices"][0]["array"], str)
 
     res2 = serializer.key_to_response(key)
-    assert np.allclose(arr, res2["choices"][0]["array"])
+    assert np.allclose(arr, res2["response"]["choices"][0]["array"])
 
 
 def test_response_to_key_numpybytes() -> None:
     """Test array serializer initialization."""
     serializer = NumpyByteSerializer()
     arr = np.random.rand(4, 4)
-    res = {"choices": [{"array": arr}]}
+    res = {"response": {"choices": [{"array": arr}]}}
     key = serializer.response_to_key(res)
     key_dct = json.loads(key)
-    assert isinstance(key_dct["choices"][0]["array"], str)
+    assert isinstance(key_dct["response"]["choices"][0]["array"], str)
 
     res2 = serializer.key_to_response(key)
-    assert np.allclose(arr, res2["choices"][0]["array"])
+    assert np.allclose(arr, res2["response"]["choices"][0]["array"])